[IMP] optimization the gitops logic

+++ src/main/java/io/choerodon/devops/app/service/impl/DevopsServiceServiceImpl.java
@@ -128,32 +128,38 @@ public class DevopsServiceServiceImpl implements DevopsServiceService {
 
     @Override
     public Boolean insertDevopsService(Long projectId, DevopsServiceReqDTO devopsServiceReqDTO) {
+        //校验环境是否链接
         envUtil.checkEnvConnection(devopsServiceReqDTO.getEnvId(), envListener);
 
         List<DevopsServiceAppInstanceE> devopsServiceAppInstanceES = new ArrayList<>();
 
+        //处理创建service对象数据
         DevopsServiceE devopsServiceE = handlerCreateService(devopsServiceReqDTO, projectId, devopsServiceAppInstanceES);
+
+        //初始化V1Service对象
         V1Service v1Service = initV1Service(
                 devopsServiceReqDTO,
                 gson.fromJson(devopsServiceE.getAnnotations(), Map.class));
-        DevopsEnvironmentE devopsEnvironmentE =
-                devopsEnviromentRepository.queryById(devopsServiceReqDTO.getEnvId());
-        UserAttrE userAttrE = userAttrRepository.queryById(TypeUtil.objToLong(GitUserNameUtil.getUserId()));
-        gitlabGroupMemberService.checkEnvProject(devopsEnvironmentE, userAttrE);
-        String path = devopsEnvironmentService.handDevopsEnvGitRepository(devopsEnvironmentE);
-        operateEnvGitLabFile(
-                TypeUtil.objToInteger(devopsEnvironmentE.getGitlabEnvProjectId()), v1Service, true, path, devopsServiceE, devopsServiceAppInstanceES, null, userAttrE);
+
+        //在gitops库处理service文件
+        operateEnvGitLabFile(v1Service, true, devopsServiceE, devopsServiceAppInstanceES, null);
         return true;
     }
 
 
     @Override
     public Boolean insertDevopsServiceByGitOps(Long projectId, DevopsServiceReqDTO devopsServiceReqDTO) {
+        //校验环境是否链接
         envUtil.checkEnvConnection(devopsServiceReqDTO.getEnvId(), envListener);
         List<DevopsServiceAppInstanceE> devopsServiceAppInstanceES = new ArrayList<>();
 
+        //处理创建service对象数据
         DevopsServiceE devopsServiceE = handlerCreateService(devopsServiceReqDTO, projectId, devopsServiceAppInstanceES);
+
+        //存储service对象到数据库
         devopsServiceE = devopsServiceRepository.insert(devopsServiceE);
+
+        //存储service和instance对象关系到数据库
         Long serviceEId = devopsServiceE.getId();
         devopsServiceAppInstanceES.parallelStream().forEach(devopsServiceAppInstanceE -> {
             devopsServiceAppInstanceE.setServiceId(serviceEId);
@@ -164,6 +170,8 @@ public class DevopsServiceServiceImpl implements DevopsServiceService {
 
 
     private DevopsServiceE handlerCreateService(DevopsServiceReqDTO devopsServiceReqDTO, Long projectId, List<DevopsServiceAppInstanceE> devopsServiceAppInstanceES) {
+
+        //校验service相关参数
         DevopsServiceValidator.checkService(devopsServiceReqDTO);
         DevopsEnvironmentE devopsEnvironmentE =
                 devopsEnviromentRepository.queryById(devopsServiceReqDTO.getEnvId());
@@ -175,6 +183,7 @@ public class DevopsServiceServiceImpl implements DevopsServiceService {
                     null);
         }
 
+        //初始化DevopsService对象
         DevopsServiceE devopsServiceE = new DevopsServiceE();
         BeanUtils.copyProperties(devopsServiceReqDTO, devopsServiceE);
         return initDevopsService(devopsServiceE, devopsServiceReqDTO, devopsServiceAppInstanceES, null);
@@ -183,7 +192,6 @@ public class DevopsServiceServiceImpl implements DevopsServiceService {
 
     private DevopsServiceE initDevopsService(DevopsServiceE devopsServiceE, DevopsServiceReqDTO devopsServiceReqDTO, List<DevopsServiceAppInstanceE> devopsServiceAppInstanceES, List<Long> beforeDevopsServiceAppInstanceES) {
         devopsServiceE.setAppId(devopsServiceReqDTO.getAppId());
-
 
         if (devopsServiceReqDTO.getLabel() != null) {
             devopsServiceE.setLabels(gson.toJson(devopsServiceReqDTO.getLabel()));
@@ -207,6 +215,7 @@ public class DevopsServiceServiceImpl implements DevopsServiceService {
     }
 
     private DevopsServiceE handlerUpdateService(DevopsServiceReqDTO devopsServiceReqDTO, DevopsServiceE devopsServiceE, List<DevopsServiceAppInstanceE> devopsServiceAppInstanceES, List<Long> beforeDevopsServiceAppInstanceES) {
+        //service参数校验
         DevopsServiceValidator.checkService(devopsServiceReqDTO);
         initDevopsServicePorts(devopsServiceReqDTO);
 
@@ -237,6 +246,7 @@ public class DevopsServiceServiceImpl implements DevopsServiceService {
             checkOptions(devopsServiceE.getEnvId(), devopsServiceReqDTO.getAppId(), null);
         }
 
+        //初始化DevopsService对象
         return initDevopsService(devopsServiceE, devopsServiceReqDTO, devopsServiceAppInstanceES, beforeDevopsServiceAppInstanceES);
 
     }
@@ -244,26 +254,25 @@ public class DevopsServiceServiceImpl implements DevopsServiceService {
     @Override
     public Boolean updateDevopsService(Long projectId, Long id,
                                        DevopsServiceReqDTO devopsServiceReqDTO) {
+        //校验环境是否链接
         envUtil.checkEnvConnection(devopsServiceReqDTO.getEnvId(), envListener);
-        DevopsEnvironmentE devopsEnvironmentE = devopsEnviromentRepository.queryById(devopsServiceReqDTO.getEnvId());
-        List<DevopsServiceAppInstanceE> devopsServiceAppInstanceES = new ArrayList<>();
 
+        //处理更新service对象数据
+        List<DevopsServiceAppInstanceE> devopsServiceAppInstanceES = new ArrayList<>();
         List<Long> beforeDevopsServiceAppInstanceES = devopsServiceInstanceRepository
                 .selectByServiceId(id).parallelStream().map(DevopsServiceAppInstanceE::getAppInstanceId).collect(Collectors.toList());
         DevopsServiceE devopsServiceE = devopsServiceRepository.query(id);
         devopsServiceE = handlerUpdateService(devopsServiceReqDTO, devopsServiceE, devopsServiceAppInstanceES, beforeDevopsServiceAppInstanceES);
-        V1Service v1Service = new V1Service();
-        if (devopsServiceE != null) {
-             v1Service = initV1Service(
+        if (devopsServiceE == null) {
+            return false;
+        } else {
+            //初始化V1Service对象
+            V1Service v1Service = initV1Service(
                     devopsServiceReqDTO,
                     gson.fromJson(devopsServiceE.getAnnotations(), Map.class));
+            //在gitops库处理service文件
+            operateEnvGitLabFile(v1Service, false, devopsServiceE, devopsServiceAppInstanceES, beforeDevopsServiceAppInstanceES);
         }
-        UserAttrE userAttrE = userAttrRepository.queryById(TypeUtil.objToLong(GitUserNameUtil.getUserId()));
-        gitlabGroupMemberService.checkEnvProject(devopsEnvironmentE, userAttrE);
-        String path = devopsEnvironmentService.handDevopsEnvGitRepository(devopsEnvironmentE);
-        operateEnvGitLabFile(
-                TypeUtil.objToInteger(devopsEnvironmentE.getGitlabEnvProjectId()), v1Service, false, path, devopsServiceE, devopsServiceAppInstanceES, beforeDevopsServiceAppInstanceES, userAttrE);
-
         return true;
     }
 
@@ -271,17 +280,22 @@ public class DevopsServiceServiceImpl implements DevopsServiceService {
     @Override
     public Boolean updateDevopsServiceByGitOps(Long projectId, Long id,
                                                DevopsServiceReqDTO devopsServiceReqDTO) {
+        //校验环境是否链接
         envUtil.checkEnvConnection(devopsServiceReqDTO.getEnvId(), envListener);
-        List<DevopsServiceAppInstanceE> devopsServiceAppInstanceES = new ArrayList<>();
 
+        //处理更新service对象数据
+        List<DevopsServiceAppInstanceE> devopsServiceAppInstanceES = new ArrayList<>();
         List<Long> beforeDevopsServiceAppInstanceES = devopsServiceInstanceRepository
                 .selectByServiceId(id).parallelStream().map(DevopsServiceAppInstanceE::getAppInstanceId).collect(Collectors.toList());
         DevopsServiceE devopsServiceE = devopsServiceRepository.query(id);
         devopsServiceE = handlerUpdateService(devopsServiceReqDTO, devopsServiceE, devopsServiceAppInstanceES, beforeDevopsServiceAppInstanceES);
-
+        if (devopsServiceE == null) {
+            return false;
+        }
+        //更新service对象到数据库
         devopsServiceRepository.update(devopsServiceE);
 
-
+        //更新service和instance关联关系数据到数据库
         beforeDevopsServiceAppInstanceES.parallelStream().forEach(instanceId ->
                 devopsServiceInstanceRepository.deleteByOptions(id, instanceId)
         );
@@ -297,20 +311,29 @@ public class DevopsServiceServiceImpl implements DevopsServiceService {
     @Override
     public void deleteDevopsService(Long id) {
         DevopsServiceE devopsServiceE = getDevopsServiceE(id);
+
+        //校验环境是否链接
         envUtil.checkEnvConnection(devopsServiceE.getEnvId(), envListener);
         devopsServiceE.setStatus(ServiceStatus.OPERATIING.getStatus());
 
 
         DevopsEnvironmentE devopsEnvironmentE = environmentRepository.queryById(devopsServiceE.getEnvId());
+
+        //判断当前容器目录下是否存在环境对应的gitops文件目录，不存在则克隆
         String path = devopsEnvironmentService.handDevopsEnvGitRepository(devopsEnvironmentE);
         UserAttrE userAttrE = userAttrRepository.queryById(TypeUtil.objToLong(GitUserNameUtil.getUserId()));
+        //检验gitops库是否存在，校验操作人是否是有gitops库的权限
         gitlabGroupMemberService.checkEnvProject(devopsEnvironmentE, userAttrE);
+
+        //查询改对象所在文件中是否含有其它对象
         DevopsEnvFileResourceE devopsEnvFileResourceE = devopsEnvFileResourceRepository
                 .queryByEnvIdAndResource(devopsEnvironmentE.getId(), id, SERVICE);
         if (devopsEnvFileResourceE == null) {
             throw new CommonException("error.fileResource.not.exist");
         }
         List<DevopsEnvFileResourceE> devopsEnvFileResourceES = devopsEnvFileResourceRepository.queryByEnvIdAndPath(devopsEnvironmentE.getId(), devopsEnvFileResourceE.getFilePath());
+
+        //如果对象所在文件只有一个对象，则直接删除文件,否则把对象从文件中去掉，更新文件
         if (devopsEnvFileResourceES.size() == 1) {
             gitlabRepository.deleteFile(
                     TypeUtil.objToInteger(devopsEnvironmentE.getGitlabEnvProjectId()),
@@ -339,7 +362,10 @@ public class DevopsServiceServiceImpl implements DevopsServiceService {
     @Override
     public void deleteDevopsServiceByGitOps(Long id) {
         DevopsServiceE devopsServiceE = getDevopsServiceE(id);
+        //校验环境是否链接
         envUtil.checkEnvConnection(devopsServiceE.getEnvId(), envListener);
+
+        //更新数据
         devopsServiceE.setStatus(ServiceStatus.OPERATIING.getStatus());
         devopsServiceRepository.update(devopsServiceE);
     }
@@ -482,25 +508,37 @@ public class DevopsServiceServiceImpl implements DevopsServiceService {
     }
 
 
-    private void operateEnvGitLabFile(Integer gitLabEnvProjectId,
-                                      V1Service service, Boolean isCreate, String path,
+    private void operateEnvGitLabFile(V1Service service, Boolean isCreate,
                                       DevopsServiceE devopsServiceE,
                                       List<DevopsServiceAppInstanceE> devopsServiceAppInstanceES,
-                                      List<Long> beforeDevopsServiceAppInstanceES,
-                                      UserAttrE userAttrE) {
+                                      List<Long> beforeDevopsServiceAppInstanceES) {
+        DevopsEnvironmentE devopsEnvironmentE =
+                devopsEnviromentRepository.queryById(devopsServiceE.getEnvId());
+        UserAttrE userAttrE = userAttrRepository.queryById(TypeUtil.objToLong(GitUserNameUtil.getUserId()));
+        //检验gitops库是否存在，校验操作人是否是有gitops库的权限
+        gitlabGroupMemberService.checkEnvProject(devopsEnvironmentE, userAttrE);
+
+        //判断当前容器目录下是否存在环境对应的gitops文件目录，不存在则克隆
+        String path = devopsEnvironmentService.handDevopsEnvGitRepository(devopsEnvironmentE);
+
+        //处理文件
         ObjectOperation<V1Service> objectOperation = new ObjectOperation<>();
         objectOperation.setType(service);
-        objectOperation.operationEnvGitlabFile("svc-" + devopsServiceE.getName(), gitLabEnvProjectId, isCreate ? "create" : "update",
+        objectOperation.operationEnvGitlabFile("svc-" + devopsServiceE.getName(), TypeUtil.objToInteger(devopsEnvironmentE.getGitlabEnvProjectId()), isCreate ? "create" : "update",
                 userAttrE.getGitlabUserId(), devopsServiceE.getId(), SERVICE, devopsServiceE.getEnvId(), path);
+
+        //操作文件成功后进行Service的数据库操作
         if (isCreate) {
             devopsServiceE = devopsServiceRepository.insert(devopsServiceE);
         } else {
             devopsServiceRepository.update(devopsServiceE);
         }
         Long serviceId = devopsServiceE.getId();
-        beforeDevopsServiceAppInstanceES.parallelStream().forEach(instanceId ->
-                devopsServiceInstanceRepository.deleteByOptions(serviceId, instanceId)
-        );
+        if (beforeDevopsServiceAppInstanceES != null) {
+            beforeDevopsServiceAppInstanceES.parallelStream().forEach(instanceId ->
+                    devopsServiceInstanceRepository.deleteByOptions(serviceId, instanceId)
+            );
+        }
         devopsServiceAppInstanceES.parallelStream().forEach(devopsServiceAppInstanceE -> {
             devopsServiceAppInstanceE.setServiceId(serviceId);
             devopsServiceInstanceRepository.insert(devopsServiceAppInstanceE);
